Add existence check to prevent overwriting config file

+++ pwss/cryptographic_algorithm/src/main/java/lib/pwss/cryptographic_algorithm/algorithm_switch/AlgorithmSwitchingInitializer.java
@@ -2,6 +2,7 @@ package lib.pwss.cryptographic_algorithm.algorithm_switch;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Properties;
@@ -16,21 +17,27 @@ public final class AlgorithmSwitchingInitializer {
         Path currentDir = Paths.get(System.getProperty("user.dir"));
         Path propertiesFilePath = currentDir.resolve(ConfigConstant.CONFIG_FILE_NAME);
 
-        Properties properties = new Properties();
+        try {
+            // Check if the file already exists
+            if (Files.exists(propertiesFilePath)) {
+                return;
+            }
 
-        try (FileOutputStream fos = new FileOutputStream(propertiesFilePath.toString())) {
-            // Add keys and values to the properties
-            properties.setProperty("1", "RSA");
-            properties.setProperty("2", "Kyber");
-            properties.setProperty("3", "Blake_2B");
+            Properties properties = new Properties();
 
-            // Save properties to file with no header comment
-            properties.store(fos, null);
+            try (FileOutputStream fos = new FileOutputStream(propertiesFilePath.toString())) {
+                // Add keys and values to the properties
+                properties.setProperty("1", "RSA");
+                properties.setProperty("2", "Kyber");
+                properties.setProperty("3", "Blake_2B");
 
-        } catch (IOException e) {
-            // Print error message if an I/O exception occurs during writing the properties file
-            System.err.println("Error writing properties to file: " + e.getMessage());
+                // Save properties to file with no header comment
+                properties.store(fos, null);
+            } catch (IOException e) {
+                return;
+            }
+        } catch (SecurityException e) {
+            return;
         }
     }
-
 }
